Fix change log deadlock bugs.

+++ src/turtle_kv/change_log/edit_offset.hpp
@@ -21,7 +21,7 @@ class WrappedInt
 {
  public:
   using Self = WrappedInt;
-  using IntT = i64;
+  using IntT = Int;
 
   //+++++++++++-+-+--+----- --- -- -  -  -   -
